server/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.29
pydantic==2.7.0
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
psycopg2-binary==2.9.9
httpx==0.27.0
email-validator==2.1.1
python-dotenv==1.0.1
slowapi==0.1.9
structlog==24.1.0
pytest==8.1.1
httpx==0.27.0
pytest-asyncio==0.23.6
PyMuPDF==1.24.2
python-docx==1.1.2
openpyxl==3.1.2
tenacity==8.2.3
celery==5.3.6
redis==5.0.3
prometheus-fastapi-instrumentator==7.0.0
